rename main component
