+++ src/App.vue
@@ -45,19 +45,19 @@
     </v-app-bar>
 
     <v-main>
-      <HelloWorld/>
+      <Main/>
     </v-main>
   </v-app>
 </template>
 
 <script>
-import HelloWorld from './components/HelloWorld';
+import Main from './components/Main';
 
 export default {
   name: 'App',
 
   components: {
-    HelloWorld,
+    Main,
   },
 
   data: () => ({
